+++ src/components/PostViewer/NonAuthModal.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useRef } from 'react';
 import styled from '@emotion/styled';
+import Modal from '@components/Modal';
 import LinkButton from '@components/NavBar/LinkButton';
 
 interface NonAuthModal {
@@ -7,29 +7,9 @@ interface NonAuthModal {
 }
 
 const NonAuthModal = ({ onClose: handleClose }: NonAuthModal) => {
-  const modalRef = useRef<null | HTMLDivElement>(null);
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Enter' || e.key === 'Escape') {
-        handleClose();
-      }
-    };
-    window.addEventListener('keydown', handleKeyDown);
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-    };
-  });
-
-  const handleClickModalContainer = (e: React.MouseEvent) => {
-    console.log('click');
-    if (e.target !== modalRef.current) {
-      handleClose();
-    }
-  };
-
   return (
-    <ModalContainer onClick={(e) => handleClickModalContainer(e)}>
-      <Modal ref={modalRef}>
+    <Modal onClose={handleClose}>
+      <>
         <Top>
           <div onClick={handleClose}>X</div>
         </Top>
@@ -47,38 +27,13 @@ const NonAuthModal = ({ onClose: handleClose }: NonAuthModal) => {
           <LinkButton to={'/login'}>로그인</LinkButton>
           <LinkButton to={'/signup'}>회원가입</LinkButton>
         </Bottom>
-      </Modal>
-    </ModalContainer>
+      </>
+    </Modal>
   );
 };
 
 export default NonAuthModal;
 
-const ModalContainer = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  backdrop-filter: blur(2px);
-  background-color: rgba(0, 0, 0, 0.2);
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-const Modal = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  width: 600px;
-  height: 400px;
-  border-radius: 50px;
-  border: 2px solid #000;
-  background-color: #fff;
-  padding: 20px;
-  box-sizing: border-box;
-`;
 const Top = styled.div`
   display: flex;
   width: 100%;
